:recycle: Extract ProcessConfig from process.go
